+++ test/unit/fuzz-tools.test.js
@@ -104,6 +104,7 @@ import * as cardioV102 from '../../lib/cardio-v102.js';
 import * as cvriskV103 from '../../lib/cvrisk-v103.js';
 import * as cardioV104 from '../../lib/cardio-v104.js';
 import * as vascularV105 from '../../lib/vascular-v105.js';
+import * as vteV106 from '../../lib/vte-v106.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -152,6 +153,7 @@ const MODULES = {
   'cvrisk-v103.js': cvriskV103,
   'cardio-v104.js': cardioV104,
   'vascular-v105.js': vascularV105,
+  'vte-v106.js': vteV106,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];